increase-plot-border
Increase the size of the border around the plot axis to fix bug #945.

Defining a fixed border size in pixels is definitely not a the best
option. A better option would be to calculate border size so that the
largest possible number "999.99" fits on the left side of the
axis. Should I try to implement that or is the larger fixed-size
border ok?

darcs-hash:20070717203144-588de-ea76eda045968b34c4406f326109966e63b917a3.gz