Calculate initial screen size by height and aspect ration
The initial size of the schematic view is calculated based on
gdk_screen_height() and gdk_screen_width(). That works fine on single
screens but on systems with two monitors (combined into one screen),
the initial window looks simply ridiculous.

To fix this, the initial windows width is calculated from the actual
height and an aspect ratio of 16:10. This will also still fit nicely
on an actual screen aspect ratio of 16:9 (HD).

Also removed an earlier call to gtk_window_set_default_size() which
had no effect whatsoever.