Transform coords "pos" to cairo_matrix_t "transform"
This is the first step for implementing the whole rotate/move properly (2nd attempt),
all based on cairo_matrix_t which can  and will be reused for proper drawing too.
Every transform regarding 0,0 on the sheet shall be described with that matrix.
To undo something the simply a stack of matrices of each item can be kept and reverted to at
will.