Hotfix: Function execution re-ordering by introducing a `g_idle_add` intermediate step, closes #130.
On loading a schematic we previously loaded all the `Wire` objects and tried to merge them,
but the view reprsentation is only created on signal callback execution which is delayed. But
before the view representation was actually created, the `Wire` itself was destroyed already
and thus left the useless view representation behind.
This is a hotfix, a not so nice workaround for a more generic problem with the current
model view approach.