This commit fixes the following issue:
https://github.com/drahnr/oregano/issues/45

by dynamically determining the sheet size from the
schematic.

Please note that this triggers the following warning
in sheet/wire-item.c:

"Failed to determine the Sheet the item is glued to"

and although there are no visible consequences, it would
be preferable to look into this.

Also, sometimes after resizing the sheet scrollbar becomes
invisible for a while (probably on large dimensions).